+++ scripts/robot_blockly_backend.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python3
+#!/usr/bin/env python
 # Software License Agreement (BSD License)
 #
 # Copyright (c) 2015, Erle Robotics LLC
@@ -38,7 +38,6 @@ import os
 import threading
 import signal
 import rosnode
-import mavros
 
 from subprocess import Popen
 from std_msgs.msg import String
@@ -46,18 +45,7 @@ from std_srvs.srv import Empty, EmptyResponse, Trigger, TriggerResponse
 from autobahn.asyncio.websocket import WebSocketServerProtocol, \
     WebSocketServerFactory
 from robot_blockly.srv import SetCurrentBlockId, SetCurrentBlockIdResponse
-from std_msgs.msg import String
 from sensor_msgs.msg import Joy
-from crab_msgs.msg import apm_imu
-from crab_msgs.msg import BodyCommand
-from crab_msgs.msg import BodyState
-from crab_msgs.msg import GaitCommand
-from crab_msgs.msg import LegIKRequest
-from crab_msgs.msg import LegJointsState
-from crab_msgs.msg import LegPositionState
-from crab_msgs.msg import LegsJointsState
-from mavros_msgs.msg import OverrideRCIn
-from mavros_msgs.srv import SetMode
 
 try:
     import asyncio
@@ -240,7 +228,7 @@ class BlocklyServerProtocol(WebSocketServerProtocol):
                             msg = Joy()
                             msg.header.stamp = rospy.Time.now()
                             rate = rospy.Rate(10)
-                               
+
                             valueAxe = 0.0
                             valueButton = 0
                             for i in range (0, 20):
@@ -263,7 +251,7 @@ class BlocklyServerProtocol(WebSocketServerProtocol):
                             rate.sleep()
 
                         elif robot.startswith('rover'):
-                            direction = robot.split('rover_')[1]   
+                            direction = robot.split('rover_')[1]
                             rospy.wait_for_service('/mavros/set_mode')
                             change_mode = rospy.ServiceProxy('/mavros/set_mode', SetMode)
                             resp1 = change_mode(custom_mode='manual')
@@ -320,11 +308,16 @@ class BlocklyServerProtocol(WebSocketServerProtocol):
         target.write("import rospy\n")
         target.write("import rosnode\n")
         target.write("import subprocess\n")
+
+        target.write("from sensor_msgs.msg import LaserScan\n")
         target.write("from std_msgs.msg import String\n")
         target.write("from std_srvs.srv import Empty, Trigger\n")
         target.write("from robot_blockly.srv import SetCurrentBlockId\n")
+        target.write("from geometry_msgs.msg import Twist\n")
         target.write("\n")
         target.write("rospy.init_node('blockly_node', anonymous=True)\n")
+        target.write("publisher = rospy.Publisher('/mobile_base/commands/velocity', Twist, queue_size=1)\n")
+        target.write("rospy.sleep(0.5)\n")
         target.write("ros_initial_nodes = rosnode.get_node_names()\n")
         target.write("\n")
         target.write("def check_status(block_id):\n")
@@ -421,7 +414,7 @@ def euler_from_quaternion(quaternion, axes='sxyz'):
     return euler_from_matrix(quaternion_matrix(quaternion), axes)''')
         target.write("\n")
         target.write("\n")
-        
+
         # Write the code that comes from blockly
         target.write(blockly_code+"\n")
         # target.write("rospy.spin()\n")